add converged to base class because current code assumes the base class always have it. as a solution in the future we can check the convergence of minibatch solve results...Then move NeuralOTProblem to its module
