+++ src/moscot/base/problems/__init__.py
@@ -2,7 +2,7 @@ from moscot.base.problems._mixins import AnalysisMixin
 from moscot.base.problems.birth_death import BirthDeathMixin, BirthDeathProblem
 from moscot.base.problems.compound_problem import BaseCompoundProblem, CompoundProblem
 from moscot.base.problems.manager import ProblemManager
-from moscot.base.problems.problem import BaseProblem, CondOTProblem, OTProblem
+from moscot.base.problems.problem import BaseProblem, OTProblem
 
 __all__ = [
     "AnalysisMixin",
@@ -13,5 +13,4 @@ __all__ = [
     "ProblemManager",
     "BaseProblem",
     "OTProblem",
-    "CondOTProblem",
 ]
