+++ src/moscot/neural/base/problems/__init__.py
@@ -1,0 +1,3 @@
+from moscot.neural.base.problems.problem import NeuralOTProblem
+
+__all__ = ["NeuralOTProblem"]
